Allow sideloading records while in flight
In some circumstances, a record may need to have its data sideloaded
while it is still in flight.

For example, when creating a parent and child in the same commit, the
child may have updated data for the parent that should be inserted into
the store. In the case of DS.RESTAdapter, the parent will still be in
flight despite having been saved first, as it will be waiting on the
child to declare the relationship change between them clean.